feat(tile-converter): save 3DNodeIndexDocument and nodepage (#2307)

+++ modules/tile-converter/src/lib/utils/file-utils.ts
@@ -1,3 +1,5 @@
+import {load} from '@loaders.gl/core';
+import {JSONLoader} from '@loaders.gl/loader-utils';
 import {promises as fs} from 'fs';
 import {isAbsolute, join} from 'path';
 import {compressFileWithGzip} from './compress-util';
@@ -40,21 +42,74 @@ export async function writeFile(
  * @param data - file content
  * @param fileName - name of output file (default: index.json)
  * @param compress - if need to compress file with gzip (default: true)
+ * @param compressList - if set - the file should be added to this list and compressed in the end of conversion
  */
 export async function writeFileForSlpk(
   path: string,
   data: string | Uint8Array | ArrayBuffer | Promise<ArrayBuffer>,
   fileName: string = 'index.json',
-  compress: boolean = true
-): Promise<string> {
+  compress: boolean = true,
+  compressList?: string[] | null
+): Promise<string | null> {
   const pathFile = await writeFile(path, data, fileName);
   if (compress) {
-    const pathGzFile = await compressFileWithGzip(pathFile);
-    // After compression, we don't need an uncompressed file
-    await removeFile(pathFile);
-    return pathGzFile;
+    if (compressList) {
+      if (!compressList.includes(pathFile)) {
+        compressList.push(pathFile);
+        return `${pathFile}.gz`;
+      } else {
+        return null;
+      }
+    } else {
+      const pathGzFile = await compressFileWithGzip(pathFile);
+      // After compression, we don't need an uncompressed file
+      await removeFile(pathFile);
+      return pathGzFile;
+    }
   }
   return pathFile;
+}
+
+/**
+ * Open json file
+ * @param path - path to the file
+ * @param fileName - file name
+ * @returns object
+ */
+export async function openJson(path: string, fileName: string): Promise<{[key: string]: any}> {
+  return new Promise((resolve, reject) => {
+    let count = 0;
+    console.log(`load ${path}/${fileName}.`); // eslint-disable-line
+    const intervalId = setInterval(() => {
+      const pathFile = join(path, fileName);
+      load(pathFile, JSONLoader)
+        .then((result) => {
+          clearInterval(intervalId);
+          resolve(result);
+        })
+        .catch(() => {
+          count++;
+          if (count > 100) {
+            clearInterval(intervalId);
+            reject(new Error(`Cannon load ${path}/${fileName}.`));
+          }
+        });
+    }, 200);
+  });
+}
+
+/**
+ * Check if the file exists
+ * @param fileName - full name of file
+ * @returns true if file exists, otherwise - false
+ */
+export async function isFileExists(fileName: string): Promise<boolean> {
+  try {
+    await fs.stat(fileName);
+    return true;
+  } catch {
+    return false;
+  }
 }
 
 /**
@@ -80,5 +135,5 @@ export function removeFile(path: string) {
  * @param filePath
  */
 export function getAbsoluteFilePath(filePath: string) {
-  return isAbsolute(filePath) ? filePath : join(process.cwd(), filePath); // eslint-disable-line no-undef
+  return isAbsolute(filePath) ? filePath : join(process.cwd(), filePath);
 }
